Move L2 backlog types from rpc/l1 to rpc/el
